rename dicecards.jsx
